[#782] Abstract provider
Problem: tzinit introduces another api for snapshots
thus we should decouple provider handling from its
implementation.

Solution: Create provider class and refactor existing code.
